\name{HMPTrees-internal}
\alias{HMPTrees-internal}
\alias{buildTree}
\alias{checkTreeValidHelp}
\alias{getBranchSizes}
\alias{getTaxaDepth}
\alias{removeUnclass}
\alias{subsetData}
\alias{traverseTree}
\alias{traverseTreeHelp}
\alias{transformHMPTreetoHMP}
\alias{transformHMPtoHMPTree}


\title{Internal Functions}
\description{These files are only called from inside other functions and are therefore not documented here.}

\author{Patricio S. La Rosa, Elena Deych, Berkley Shands, William D. Shannon}

\keyword{internal}