\name{HMPTrees-package}
\alias{HMPTrees-package}
\alias{HMPTrees}
\docType{package}
\title{Statistical Object Oriented Data Analysis of RDP-based Taxonomic trees from Human Microbiome Data: Modeling, Visualization, and Two-Group Comparison}

\description{In this package, we introduce Object Oriented Data Analysis (OODA) methods to analyze Human Microbiome taxonomic 
trees directly. We provide tools to model, compare, and visualize populations of taxonomic tree objects. In particular, 
using a unimodal probability measure in the set of trees, we provide procedures to compute the MLE of the central tree and the 
LRT statistics for comparing the distributions of two metagenomic populations. Our procedures have been developed to support parallel 
computations for comparing two groups.}

\details{
\tabular{ll}{
Package: \tab HMPTrees\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-11-16\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Patricio S. La Rosa, Elena Deych, Berkley Shands, William D. Shannon
Maintainer: Berkley Shands <bshands@dom.wustl.edu>
}

\keyword{ package }
\keyword{ internal }