\name{HMPTrees-internal}
\alias{HMPTrees-internal}
\alias{checkValidHelp}
\alias{traverseHelp}
\alias{helpRelations}
\alias{rowMatch}
\alias{getTopXPhylum}
\alias{getBranchSizes}
\alias{getRelations}
\alias{getTaxaDepth}
\alias{traverseTree}

\title{Internal Functions}
\description{These files are for internal use only.}
\author{Patricio S. La Rosa, Elena Deych, Berkley Shands, William D. Shannon}

\keyword{internal}